requirements.txt
pyyaml>=6.0.1
openai>=1.54.0
pypdf>=5.1.0
pandas>=2.2.0
numpy>=1.26.0

